Schema updated

+++ src/main/java/io/kokilaw/banking/repository/CurrencyRepository.java
@@ -8,6 +8,6 @@ import java.util.Optional;
 /**
  * Created by kokilaw on 2022-08-11
  */
-public interface CurrencyRepository extends JpaRepository<Currency, Long> {
+public interface CurrencyRepository extends JpaRepository<Currency, Integer> {
     Optional<Currency> findCurrencyByCurrencyCode(String currencyCode);
 }
